Move cookie banner into seperate file
